Add face-capturing functionality
This program is a carbon-copy of the facial recognition from project X0, except when it detects a face, it saves a cropped version of that image to a file. (Cropped around the face it detected.)

Currently, this program does not have functionality to specify the subject of the images for model-training purposes.