Add ability to label data.
The program now takes the first command-line argument as a label for the images it saves. Files are saved with the following format: "datasets/[label]/face_[label]_[image number]" where "[label]" is the specified label, and "[image_number]" is the value denoting which image it is. Currently, this value starts at 0 every time the program runs, so subsequent executions of this script will overwrite any previous ones' images.